+++ Assets/OxGKit/Utilities/Scripts/Runtime/Cacher/Interface/IRemoveCacheHandler.cs
@@ -1,0 +1,7 @@
+namespace OxGKit.Utilities.Cacher
+{
+    public interface IRemoveCacheHandler<TKey, TValue>
+    {
+        void RemoveCache(TKey key, TValue value);
+    }
+}